refactor: separate chain watcher
